Added support for MultiLine, MultiPolygon and MultiPoint in drawtools (#1883)
